third_party/blink/web_tests/wpt_internal/css/css-mixins/basic.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Mixins: Basic test</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
      @mixin --m1 {
        .cls {
          color: green;
        }
      }
      @mixin invalid-name {
        .cls {
          color: red;
        }
      }
      @mixin --empty {}
      div {
        @apply --m1;
        @apply invalid-name;
        @apply --empty;
      }
    </style>
  </head>
  <body>
    <div><div class="cls" id="target">This text should be green.</div></div>
  <script>
    test(() => {
      let target = document.getElementById('target');
      assert_equals(getComputedStyle(target).color, 'rgb(0, 128, 0)');
    });
  </script>
  </body>
</html>
